Refactor product-summary-image to TypeScript
