--- react/ProductSummaryImage.js
@@ -1,3 +1,0 @@
-import ProductImage from './components/ProductSummaryImage/ProductImage'
-
-export default ProductImage
