+++ react/__mocks__/vtex.product-summary-context/ProductSummaryContext.js
@@ -1,15 +1,21 @@
 import React, { createContext, useContext } from 'react'
 
-const ProductSummaryContext = createContext({})
+const Context = createContext({})
 
 export const ProductSummaryProvider = ({ product, ...rest }) => {
-  return <ProductSummaryContext.Provider value={{ product }} {...rest} />
+  return <Context.Provider value={{ product }} {...rest} />
 }
 
 export const useProductSummary = () => {
-  return useContext(ProductSummaryContext)
+  return useContext(Context)
 }
 
 export const useProductSummaryDispatch = () => {
   return () => {}
+}
+
+export const ProductSummaryContext = {
+  ProductSummaryProvider,
+  useProductSummary,
+  useProductSummaryDispatch,
 }
